requirements.txt
# Full dependencies — for MAESTRO (indexing + matching + SageMaker endpoint)
openai>=1.0.0
faiss-cpu>=1.7.4
boto3>=1.28.0
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
